app/requirements.txt
streamlit
pandas
aiohttp
openai
openpyxl
streamlit-option-menu
google-generativeai
requests